fix: move logic to new sub command
